Add stock-indicator.py
A Gnome Unity appindicator that displays the current price of a given stock.
It uses the Yahoo Finance stock API to fetch the price for one stock, and
displays that price in the Notification area (the grey menu bar at the top
of the screen).

Based on the Gmail appindicator from
http://conjurecode.com/create-indicator-applet-for-ubuntu-unity-with-python/

Signed-off-by: Timur Tabi <timur@tabi.org>